requirements.txt
certifi==2018.11.29
chardet==3.0.4
datadog==0.26.0
decorator==4.3.2
Django==2.1.7
django-appconf==1.0.2
djangorestframework==3.9.1
idna==2.8
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
